peg: repair src URL
The webserver doesn't map // → /. The project's homepage does have (and
does require) a trailing slash (without redirection). So, we've patched
it in this way. An alternative way would be to fold all repeated /s into
a single /. But, this is cheaper to evaluate if a bit awkward to
read/parse as a human.

```
for p in /software/peg /software/peg/ /software/peg//0.1.20.tar.gz /software/peg/0.1.20.tar.gz; do
 curl --silent --write-out '%{http_code}\n' https://piumarta.com$p -o /dev/null;
done

301
200
404
404
```